refactor: 로그인 실패 시 예외 catch 지점 변경 및 테스트 이동
